# Optimized "computing pnes" in classical calculations (3x in the common case)