--- a/debian/changelog
+++ b/debian/changelog
@@ -1,3 +1,6 @@
+  [Michele Simionato]
+  * Optimized "computing pnes" in classical calculations (3x in the common case)
+
   [Nicolas Schmid]
   * Fixed a regression in the ShakeMap to_gmfs code, a forgotten sigma^2/2 term