Here are the figures for performance.zip on my machine:
```
# before
| calc_39754, maxmem=10.0 GB | time_sec | memory_mb | counts    |
|----------------------------+----------+-----------+-----------|
| total classical            | 4_065    | 78.8398   | 24        |
| get_poes                   | 2_331    | 0.0       | 1_066_484 |
| computing mean_std         | 712.5    | 0.0       | 21_330    |
| composing pnes             | 677.6    | 0.0       | 1_066_484 |
| ClassicalCalculator.run    | 543.8    | 62.8125   | 1         |
| planar contexts            | 224.9    | 0.0       | 342_760   |
| total postclassical        | 22.2600  | 1.4883    | 10        |

# after, speedup 2.8x in "computing pnes", in total down to 8 minutes from 9 minutes
| calc_39753, maxmem=12.1 GB | time_sec | memory_mb | counts    |
|----------------------------+----------+-----------+-----------|
| total classical            | 3_522    | 75.9023   | 24        |
| get_poes                   | 2_268    | 0.0       | 1_066_484 |
| computing mean_std         | 687.7    | 0.0       | 21_330    |
| ClassicalCalculator.run    | 476.8    | 62.5781   | 1         |
| composing pnes             | 240.8    | 0.0       | 1_066_484 |
| planar contexts            | 220.3    | 0.0       | 342_760   |
| total postclassical        | 21.8133  | 1.3750    | 10        |
```